moved harmony patches out of BasePlugin.cs and gave each patch a class-specific file
