+++ ChebsNecromancy/BasePlugin.cs
@@ -22,9 +22,6 @@ using Jotunn.Utils;
 using UnityEngine;
 using Paths = BepInEx.Paths;
 
-// ReSharper disable InconsistentNaming
-// ReSharper disable UnusedParameter.Local
-
 namespace ChebsNecromancy
 {
     [BepInPlugin(PluginGuid, PluginName, PluginVersion)]
@@ -395,452 +392,5 @@ namespace ChebsNecromancy
             spectralShroudItem.DoOnUpdate();
         }
     }
-
-    #region HarmonyPatches
-
-    // Harmony patching is very sensitive regarding parameter names. Everything in this region should be hand crafted
-    // and not touched by well-meaning but clueless IDE optimizations.
-    // eg.
-    // __instance MUST be named with exactly two underscores.
-    // ___m_drops MUST be named with exactly three underscores.
-    //
-    // This is because all of this has a special meaning to Harmony.
-    
-    [HarmonyPatch(typeof(CharacterDrop), "GenerateDropList")]
-    class CharacterDrop_Patches
-    {
-        [HarmonyPrefix]
-#pragma warning disable IDE0051 // Remove unused private members
-        static void AddBonesToDropList(ref List<CharacterDrop.Drop> ___m_drops)
-#pragma warning restore IDE0051 // Remove unused private members
-        {
-            if (SkeletonWand.BoneFragmentsDroppedAmountMin.Value >= 0
-                && SkeletonWand.BoneFragmentsDroppedAmountMax.Value > 0)
-            {
-                CharacterDrop.Drop bones = new()
-                {
-                    m_prefab = ZNetScene.instance.GetPrefab("BoneFragments"),
-                    m_onePerPlayer = true,
-                    m_amountMin = SkeletonWand.BoneFragmentsDroppedAmountMin.Value,
-                    m_amountMax = SkeletonWand.BoneFragmentsDroppedAmountMax.Value,
-                    m_chance = 1f
-                };
-                ___m_drops.Add(bones);
-            }
-        }
-    }
-
-    [HarmonyPatch(typeof(Piece))]
-    static class ChebGonaz_PiecePatch
-    {
-        [HarmonyPostfix]
-        [HarmonyPatch(nameof(Piece.Awake))]
-        static void AwakePostfix(ref Piece __instance)
-        {
-            if (__instance.name.StartsWith("ChebGonaz"))
-            {
-                if (__instance.name.Contains("SpiritPylon"))
-                {
-                    if (__instance.GetComponent<SpiritPylon>() == null)
-                    {
-                        __instance.gameObject.AddComponent<SpiritPylon>();
-                    }
-                }
-                else if (__instance.name.Contains("RefuelerPylon"))
-                {
-                    if (__instance.GetComponent<RefuelerPylon>() == null)
-                    {
-                        __instance.gameObject.AddComponent<RefuelerPylon>();
-                    }
-                }
-                else if (__instance.name.Contains("NeckroGathererPylon"))
-                {
-                    if (__instance.GetComponent<NeckroGathererPylon>() == null)
-                    {
-                        __instance.gameObject.AddComponent<NeckroGathererPylon>();
-                    }
-                }
-                else if (__instance.name.Contains("BatBeacon"))
-                {
-                    if (__instance.GetComponent<BatBeacon>() == null)
-                    {
-                        __instance.gameObject.AddComponent<BatBeacon>();
-                    }
-                }
-            }
-        }
-    }
-
-    [HarmonyPatch(typeof(MonsterAI))]
-    static class FriendlySkeletonPatch
-    {
-        [HarmonyPostfix]
-        [HarmonyPatch(nameof(MonsterAI.Awake))]
-        static void AwakePostfix(ref Character __instance)
-        {
-            if (__instance.name.StartsWith("ChebGonaz"))
-            {
-                if (__instance.name.Contains("Wraith"))
-                {
-                    __instance.gameObject.AddComponent<GuardianWraithMinion>();
-                }
-                else
-                if (__instance.name.Contains("SpiritPylonGhost") && __instance.GetComponent<SpiritPylonGhostMinion>() == null)
-                {
-                    __instance.gameObject.AddComponent<SpiritPylonGhostMinion>();
-                }
-                else
-                {
-                    if (__instance.GetComponent<UndeadMinion>() == null)
-                    {
-                        if (__instance.name.Contains("PoisonSkeleton"))
-                        {
-                            __instance.gameObject.AddComponent<PoisonSkeletonMinion>();
-                        }
-                        else if (__instance.name.Contains("Skeleton"))
-                        {
-                            __instance.gameObject.AddComponent<SkeletonMinion>();
-                        }
-                        else if (__instance.name.Contains("Draugr"))
-                        {
-                            __instance.gameObject.AddComponent<DraugrMinion>();
-                        }
-                        else if (__instance.name.Contains("Neckro"))
-                        {
-                            __instance.gameObject.AddComponent<NeckroGathererMinion>();
-                        }
-                        else if (__instance.name.Contains("Bat"))
-                        {
-                            __instance.gameObject.AddComponent<BatBeaconBatMinion>();
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    [HarmonyPatch(typeof(WearNTear), "RPC_Damage")]
-    static class ArrowImpactPatch
-    {
-        // stop minions from damaging player structures
-#pragma warning disable IDE0051 // Remove unused private members
-        static void Prefix(ref HitData hit, Piece ___m_piece)
-#pragma warning restore IDE0051 // Remove unused private members
-        {
-            if (hit != null)
-            {
-                Character attacker = hit.GetAttacker();
-                if (attacker != null 
-                    && attacker.TryGetComponent(out UndeadMinion _))
-                {
-                    if (___m_piece.IsPlacedByPlayer())
-                    {
-                        hit.m_damage.m_damage = 0f;
-                        hit.m_damage.m_blunt = 0f;
-                        hit.m_damage.m_slash = 0f;
-                        hit.m_damage.m_pierce = 0f;
-                        hit.m_damage.m_chop = 0f;
-                        hit.m_damage.m_pickaxe = 0f;
-                        hit.m_damage.m_fire = 0f;
-                        hit.m_damage.m_frost = 0f;
-                        hit.m_damage.m_lightning = 0f;
-                        hit.m_damage.m_poison = 0f;
-                        hit.m_damage.m_spirit = 0f;
-                    }
-                }
-            }
-        }
-    }
-
-    [HarmonyPatch(typeof(CharacterDrop), "OnDeath")]
-    static class OnDeathDropPatch
-    {
-#pragma warning disable IDE0051 // Remove unused private members
-        static bool Prefix(CharacterDrop __instance)
-#pragma warning restore IDE0051 // Remove unused private members
-        {
-            // Although Container component is on the Neckro, its OnDestroyed
-            // isn't called on the death of the creature. So instead, implement
-            // its same functionality in the creature's OnDeath instead.
-            if (__instance.TryGetComponent(out NeckroGathererMinion necroNeck))
-            {
-                if (__instance.TryGetComponent(out Container container))
-                {
-                    container.DropAllItems(container.m_destroyedLootPrefab);
-                    return false; // deny base method completion
-                }
-            }
-
-            // For all other minions, check if they're supposed to be dropping
-            // items and whether tehse should be packed into a crate or not.
-            // We don't want ppls surtling cores and things to be claimed by davey jones
-            else if (__instance.TryGetComponent(out UndeadMinion undeadMinion))
-            {
-                void PackDropsIntoCrate()
-                {
-                    // use vanilla cargo crate -> same as a karve/longboat drops
-                    GameObject cratePrefab = ZNetScene.instance.GetPrefab("CargoCrate");
-                    if (cratePrefab != null)
-                    {
-                        // warning: we mustn't ever exceed the maximum storage capacity
-                        // of the crate. Not a problem right now, but could be in the future
-                        // if the ingredients exceed 4. Right now, can only be 3, so it's fine.
-                        // eg. bones, meat, ingot (draugr) OR bones, ingot, surtling core (skele)
-                        Inventory inv =
-                            GameObject.Instantiate(cratePrefab, __instance.transform.position + Vector3.up, Quaternion.identity)
-                            .GetComponent<Container>()
-                            .GetInventory();
-                        __instance.m_drops.ForEach(drop => inv.AddItem(drop.m_prefab, drop.m_amountMax));
-                    }
-                }
-
-                if (undeadMinion is SkeletonMinion
-                    && SkeletonMinion.DropOnDeath.Value != UndeadMinion.DropType.Nothing
-                    && SkeletonMinion.PackDropItemsIntoCargoCrate.Value)
-                {
-                    PackDropsIntoCrate();
-                    return false; // deny base method completion
-                }
-                if (undeadMinion is DraugrMinion
-                    && DraugrMinion.DropOnDeath.Value != UndeadMinion.DropType.Nothing
-                    && DraugrMinion.PackDropItemsIntoCargoCrate.Value)
-                {
-                    PackDropsIntoCrate();
-                    return false; // deny base method completion
-                }
-            }
-
-            return true; // permit base method to complete
-        }
-    }
-
-    [HarmonyPatch(typeof(Character), "RPC_Damage")]
-    static class CharacterGetDamageModifiersPatch
-    {
-        // here we basically have to rewrite the entire RPC_Damage verbatim
-        // except including the GetBodyArmor that is usually only kept
-        // for players and omitted for NPCs. We also discard the durability
-        // stuff cuz that doesn't matter for NPCs.
-        //
-        // I also pruned some player stuff out.
-#pragma warning disable IDE0051 // Remove unused private members
-#pragma warning disable IDE0060 // Remove unused parameter
-        static bool Prefix(ref long sender, ref HitData hit, Character __instance)
-#pragma warning restore IDE0060 // Remove unused parameter
-#pragma warning restore IDE0051 // Remove unused private members
-        {
-            if (__instance.TryGetComponent(out UndeadMinion minion)
-                && (minion is SkeletonMinion || minion is DraugrMinion))
-            {
-                if (!__instance.m_nview.IsOwner()
-                    || __instance.GetHealth() <= 0f
-                    || __instance.IsDead()
-                    || __instance.IsTeleporting() 
-                    || __instance.InCutscene() 
-                    || (hit.m_dodgeable && __instance.IsDodgeInvincible()))
-                {
-                    return false; // deny base method completion
-                }
-                Character attacker = hit.GetAttacker();
-                if (hit.HaveAttacker() && attacker == null)
-                {
-                    return false; // deny base method completion
-                }
-                if (attacker != null && !attacker.IsPlayer())
-                {
-                    float difficultyDamageScalePlayer = Game.instance.GetDifficultyDamageScalePlayer(__instance.transform.position);
-                    hit.ApplyModifier(difficultyDamageScalePlayer);
-                }
-                __instance.m_seman.OnDamaged(hit, attacker);
-                if (__instance.m_baseAI !=null 
-                    && __instance.m_baseAI.IsAggravatable() 
-                    && !__instance.m_baseAI.IsAggravated())
-                {
-                    BaseAI.AggravateAllInArea(__instance.transform.position, 20f, BaseAI.AggravatedReason.Damage);
-                }
-                if (__instance.m_baseAI != null 
-                    && !__instance.m_baseAI.IsAlerted() 
-                    && hit.m_backstabBonus > 1f 
-                    && Time.time - __instance.m_backstabTime > 300f)
-                {
-                    __instance.m_backstabTime = Time.time;
-                    hit.ApplyModifier(hit.m_backstabBonus);
-                    __instance.m_backstabHitEffects.Create(hit.m_point, Quaternion.identity, __instance.transform);
-                }
-                if (__instance.IsStaggering() && !__instance.IsPlayer())
-                {
-                    hit.ApplyModifier(2f);
-                    __instance.m_critHitEffects.Create(hit.m_point, Quaternion.identity, __instance.transform);
-                }
-                if (hit.m_blockable && __instance.IsBlocking())
-                {
-                    __instance.BlockAttack(hit, attacker);
-                }
-                __instance.ApplyPushback(hit);
-                if (!string.IsNullOrEmpty(hit.m_statusEffect))
-                {
-                    StatusEffect statusEffect = __instance.m_seman.GetStatusEffect(hit.m_statusEffect);
-                    if (statusEffect == null)
-                    {
-                        statusEffect = __instance.m_seman.AddStatusEffect(
-                            hit.m_statusEffect, 
-                            false,
-                            hit.m_itemLevel, 
-                            hit.m_skillLevel);
-                    }
-                    else
-                    {
-                        statusEffect.ResetTime();
-                        statusEffect.SetLevel(hit.m_itemLevel, hit.m_skillLevel);
-                    }
-                    if (statusEffect != null && attacker != null)
-                    {
-                        statusEffect.SetAttacker(attacker);
-                    }
-                }
-                WeakSpot weakSpot = __instance.GetWeakSpot(hit.m_weakSpot);
-                if (weakSpot != null)
-                {
-                    ZLog.Log($"HIT Weakspot: {weakSpot.gameObject.name}");
-                }
-                HitData.DamageModifiers damageModifiers = __instance.GetDamageModifiers(weakSpot);
-                hit.ApplyResistance(damageModifiers, out var significantModifier);
-                // THIS is what we wrote all the code above for...
-                //
-                // GetBodyArmor should work, but doesn't. So we tally it up
-                // ourselves.
-                //
-                //float bodyArmor = __instance.GetBodyArmor();
-                float bodyArmor = 0f;
-
-                if (__instance.TryGetComponent(out Humanoid humanoid))
-                {
-                    bodyArmor += humanoid.m_chestItem != null
-                        ? humanoid.m_chestItem.m_shared.m_armor : 0;
-
-                    bodyArmor += humanoid.m_legItem != null
-                        ? humanoid.m_legItem.m_shared.m_armor : 0;
-
-                    bodyArmor += humanoid.m_helmetItem != null
-                        ? humanoid.m_helmetItem.m_shared.m_armor : 0;
-                }
-                bodyArmor *= SkeletonWand.SkeletonArmorValueMultiplier.Value;
-                hit.ApplyArmor(bodyArmor);
-                //Jotunn.Logger.LogInfo($"{__instance.name} applied body armor {bodyArmor}");
-                // // //
-                float poison = hit.m_damage.m_poison;
-                float fire = hit.m_damage.m_fire;
-                float spirit = hit.m_damage.m_spirit;
-                hit.m_damage.m_poison = 0f;
-                hit.m_damage.m_fire = 0f;
-                hit.m_damage.m_spirit = 0f;
-                __instance.ApplyDamage(hit, true, true, significantModifier);
-                __instance.AddFireDamage(fire);
-                __instance.AddSpiritDamage(spirit);
-                __instance.AddPoisonDamage(poison);
-                __instance.AddFrostDamage(hit.m_damage.m_frost);
-                __instance.AddLightningDamage(hit.m_damage.m_lightning);
-                return false; // deny base method completion
-            }
-            return true; // permit base method to complete
-        }
-    }
-
-    [HarmonyPatch(typeof(Aoe), "OnHit")]
-    static class SharpStakesMinionPatch
-    {
-        // bool OnHit(Collider collider, Vector3 hitPoint)
-#pragma warning disable IDE0051 // Remove unused private members
-#pragma warning disable IDE0060 // Remove unused parameter
-        static bool Prefix(Collider collider, Vector3 hitPoint, Aoe __instance)
-#pragma warning restore IDE0060 // Remove unused parameter
-#pragma warning restore IDE0051 // Remove unused private members
-        {
-            if (collider.TryGetComponent(out UndeadMinion _))
-            {
-                Piece piece = __instance.GetComponentInParent<Piece>();
-                if (piece != null && piece.IsPlacedByPlayer())
-                {
-                    // stop minion from receiving damage from stakes placed
-                    // by a player
-                    return false; // deny base method completion
-                }
-            }
-            return true; // permit base method to complete
-        }
-    }
-
-    [HarmonyPatch(typeof(Tameable), "Interact")]
-    static class TameablePatch1
-    {
-        [HarmonyPrefix]
-#pragma warning disable IDE0051 // Remove unused private members
-#pragma warning disable IDE0060 // Remove unused parameter
-        static bool InteractPrefix(Humanoid user, bool hold, bool alt, Tameable __instance)
-#pragma warning restore IDE0060 // Remove unused parameter
-#pragma warning restore IDE0051 // Remove unused private members
-        {
-            // Stop players that aren't the owner of a minion from interacting
-            // with it. Also call UndeadMinion wait/follow methods to
-            // properly update the ZDO with the waiting position.
-            if (__instance.TryGetComponent(out UndeadMinion undeadMinion)
-                && user.TryGetComponent(out Player player))
-            {
-                if (!undeadMinion.BelongsToPlayer(player.GetPlayerName()))
-                {
-                    user.Message(MessageHud.MessageType.Center, "$chebgonaz_notyourminion");
-                    return false; // deny base method completion
-                }
-
-                if (!UndeadMinion.Commandable.Value)
-                {
-                    return false; // deny base method completion
-                }
-
-                // use the minion methods to ensure the ZDO is updated
-                if (__instance.TryGetComponent(out MonsterAI monsterAI))
-                {
-                    if (monsterAI.GetFollowTarget() == player.gameObject)
-                    {
-                        user.Message(MessageHud.MessageType.Center, "$friendlyskeletonwand_skeletonwaiting");
-                        undeadMinion.Wait(player.transform.position);
-                    }
-                    else
-                    {
-                        user.Message(MessageHud.MessageType.Center, "$friendlyskeletonwand_skeletonfollowing");
-                        undeadMinion.Follow(player.gameObject);
-                    }
-                    return false; // deny base method completion
-                }
-            }
-
-            return true; // permit base method to complete
-        }
-    }
-
-    [HarmonyPatch(typeof(Tameable))]
-    static class TameablePatch2
-    {
-        [HarmonyPatch(nameof(Tameable.GetHoverText))]
-        [HarmonyPrefix]
-        static bool Prefix(Tameable __instance, ref string __result)
-        {
-            if (__instance.m_nview.IsValid()
-                && __instance.m_commandable
-                && __instance.TryGetComponent(out UndeadMinion _)
-                && __instance.TryGetComponent(out MonsterAI monsterAI)
-                && Player.m_localPlayer != null)
-            {
-                __result = monsterAI.GetFollowTarget() == Player.m_localPlayer.gameObject
-                    ? Localization.instance.Localize("$chebgonaz_wait")
-                    : Localization.instance.Localize("$chebgonaz_follow");
-                return false; // deny base method completion
-            }
-
-            return true; // allow base method completion
-        }
-    }
-
-    #endregion
 }
