+++ ChebsNecromancy/ChebsNecromancy.csproj
@@ -135,6 +135,13 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Common\ChebsRecipe.cs" />
+    <Compile Include="Patches\AoePatches.cs" />
+    <Compile Include="Patches\CharacterDropPatches.cs" />
+    <Compile Include="Patches\CharacterPatches.cs" />
+    <Compile Include="Patches\MonsterAIPatches.cs" />
+    <Compile Include="Patches\PiecePatches.cs" />
+    <Compile Include="Patches\TameablePatches.cs" />
+    <Compile Include="Patches\WearNTearPatches.cs" />
     <Compile Include="Properties\IgnoreAccessModifiers.cs" />
     <Compile Include="BasePlugin.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
